feat: mockNvxDevice respects the `mode` property
In testing, it is sometimes beneficial to allow for having mock devices
that allow for checking routing behavior even without real hardware. The
Mock NVX Device will now respect the same `mode` property that the real
devices do, creating all the tielines in the central router as
appropriate.

In addition, the `SecondaryAudio` flag has been added where appropriate
to allow for the correct behavior when attempting to route Secondary
Audio in an NVX switch fabric.